# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "KubeFin Cost Analyzer API.",
        "title": "KubeFin API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/resources/clusters/{cluster_id}/cpu": {
            "get": {
                "description": "Get specific cluster CPU resources metrics",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Get specific cluster CPU resources metrics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Id",
                        "name": "cluster_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The start time to query",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The end time to query",
                        "name": "endTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The step seconds of the data to return",
                        "name": "stepSeconds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterResourceMetrics"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.StatusError"
                        }
                    }
                }
            }
        },
        "/resources/clusters/{cluster_id}/memory": {
            "get": {
                "description": "Get specific cluster memory resources metrics",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Get specific cluster memory resources metrics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Id",
                        "name": "cluster_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The start time to query",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The end time to query",
                        "name": "endTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The step seconds of the data to return",
                        "name": "stepSeconds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterResourceMetrics"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.StatusError"
                        }
                    }
                }
            }
        },
        "/resources/clusters/{cluster_id}/summary": {
            "get": {
                "description": "Get specific cluster resources summary in current two month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Get specific cluster resources summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Id",
                        "name": "cluster_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterResourcesSummary"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.StatusError"
                        }
                    }
                }
            }
        },
        "/resources/summary": {
            "get": {
                "description": "Get all clusters resources summary in current two month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Get all clusters resources summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterResourcesSummaryList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.StatusError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_kubefin_kubefin_pkg_api.ClusterResourceMetrics": {
            "type": "object",
            "properties": {
                "clusterId": {
                    "type": "string"
                },
                "resourceAvailableValues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SamplePair"
                    }
                },
                "resourceRequestValues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SamplePair"
                    }
                },
                "resourceSystemTakenValues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SamplePair"
                    }
                },
                "resourceTotalValues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SamplePair"
                    }
                },
                "resourceType": {
                    "type": "string"
                },
                "resourceUsageValues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SamplePair"
                    }
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.ClusterResourcesSummary": {
            "type": "object",
            "properties": {
                "cloudProvider": {
                    "type": "string"
                },
                "clusterActiveTime": {
                    "description": "ClusterActiveTime shows the cluster active time in seconds",
                    "type": "number"
                },
                "clusterConnectionSate": {
                    "description": "ClusterConnectionSate can be running/connect_failed",
                    "type": "string"
                },
                "clusterId": {
                    "type": "string"
                },
                "clusterName": {
                    "type": "string"
                },
                "clusterRegion": {
                    "type": "string"
                },
                "connectionTime": {
                    "description": "ConnectionTime shows the time the cluster connected",
                    "type": "integer"
                },
                "cpuCoreAvailable": {
                    "description": "CPUCoreAvailable means all nodes' available cpu core",
                    "type": "number"
                },
                "cpuCoreRequest": {
                    "description": "CPUCoreRequest means all pods' cpu core request",
                    "type": "number"
                },
                "cpuCoreSystemTaken": {
                    "description": "CPUCoreSystemTaken means the cpu taken by system",
                    "type": "number"
                },
                "cpuCoreTotal": {
                    "description": "CPUCoreTotal means all nodes' cpu core",
                    "type": "number"
                },
                "cpuCoreUsage": {
                    "description": "CPUCoreUsage means all pods' cpu core usage",
                    "type": "number"
                },
                "fallbackBillingNodeNumbersCurrent": {
                    "type": "integer"
                },
                "lastActiveTime": {
                    "type": "integer"
                },
                "nodeNumbersCurrent": {
                    "type": "integer"
                },
                "onDemandBillingNodeNumbersCurrent": {
                    "type": "integer"
                },
                "periodBillingNodeNumbersCurrent": {
                    "type": "integer"
                },
                "podScheduledCurrent": {
                    "type": "integer"
                },
                "podTotalCurrent": {
                    "type": "integer"
                },
                "podUnscheduledCurrent": {
                    "type": "integer"
                },
                "ramGiBAvailable": {
                    "description": "RAMGiBAvailable means all nodes' available ram GiB",
                    "type": "number"
                },
                "ramGiBRequest": {
                    "description": "RAMGiBRequest all pods' ram GiB request",
                    "type": "number"
                },
                "ramGiBSystemTaken": {
                    "description": "RAMGiBSystemTaken means the rm taken by system",
                    "type": "number"
                },
                "ramGiBTotal": {
                    "description": "RAMGiBTotal means all nodes' ram GiB",
                    "type": "number"
                },
                "ramGiBUsage": {
                    "description": "RAMGiBUsage means all pods' ram GiB usage",
                    "type": "number"
                },
                "spotBillingNodeNumbersCurrent": {
                    "type": "integer"
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.ClusterResourcesSummaryList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterResourcesSummary"
                    }
                }
            }
        },
...
        "github_com_kubefin_kubefin_pkg_api.StatusError": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.SamplePair": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}